+++ Business/Handlers/Assets/Commands/UpdateAssetCommand.cs
@@ -26,7 +26,7 @@ namespace Business.Handlers.Assets.Commands
         public int Id { get; set; }
         public int AssetTypeId { get; set; }
         public string Name { get; set; }
-        public int Piece { get; set; }
+        public decimal Piece { get; set; }
         public bool IsDebt { get; set; } = false;
 
         public class UpdateAssetCommandHandler : IRequestHandler<UpdateAssetCommand, IResult>
